+++ .jungle-kit/scripts/smudge-local.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-# Jungle Kit: Passthrough smudge filter
-cat